+++ Client/src/views/Detail/Detail.jsx
@@ -30,7 +30,7 @@ const Detail = () => {
 
     return (
         <div>
-            <Link to="/home">
+            <Link to="/">
                 <button><span>Volver</span></button>
             </Link>
             {
