+++ Client/src/views/ProductList/ProductList.module.css
@@ -15,3 +15,8 @@
    max-width: 100%;
    height: 25rem;
 }
+
+.productLink {
+   /* min-width: 100%; */
+   
+}
